+++ Core/Models/Settings/App.cs
@@ -2,18 +2,11 @@
 {
     public class App
     {
-        #region Private Fields
-
-        private const int HeightDefault = 800;
-        private const int WidthDefault = 1200;
-
-        #endregion Private Fields
-
         #region Public Properties
 
-        public int Height { get; set; } = HeightDefault;
+        public int Height { get; set; } = Constants.AppHeightDefault;
 
-        public int Width { get; set; } = WidthDefault;
+        public int Width { get; set; } = Constants.AppWidthDefault;
 
         public string WindowState { get; set; } = nameof(Avalonia.Controls.WindowState.Maximized);