+++ Core/Models/Settings/Detection.cs
@@ -2,23 +2,17 @@
 {
     public class Detection
     {
-        #region Private Fields
-
-        private const int ThresholdDetectingDefault = 90;
-        private const int ThresholdMatchingDefault = 40;
-        private const int WaitingDurationDefault = 100;
-
-        #endregion Private Fields
-
         #region Public Properties
 
         public bool NoRecognition { get; set; }
 
-        public int ThresholdDetecting { get; set; } = ThresholdDetectingDefault;
+        public float Rotation { get; set; }
 
-        public int ThresholdMatching { get; set; } = ThresholdMatchingDefault;
+        public int ThresholdDetecting { get; set; } = Constants.DetectionThresholdDefault;
 
-        public int WaitingDuration { get; set; } = WaitingDurationDefault;
+        public int ThresholdMatching { get; set; } = Constants.DetectionMatchingDefault;
+
+        public int WaitingDuration { get; set; } = Constants.DetectionWaitingDefault;
 
         #endregion Public Properties
     }